requirements.txt
python-twitter == 3.5